moved all the sharelist logic from the managelist component into a new component: ShareList
